chore

+++ src/wall_terminal/platformio_install_unilibc.sh
@@ -11,8 +11,12 @@ git clone $REPO_URL tmp
 mkdir -p lib
 
 for i in "${!libs[@]}"; do
-	rm -rf "lib/ul_${libs[$i]}"
-	mv "tmp/lib/ul_${libs[$i]}" lib
+
+	# Do not overwrite the `ul_configs` lib but, if does not exist, use the default one.
+	if [[ "${libs[$i]}" != "configs" || ( "${libs[$i]}" == "configs" && ! -d "lib/ul_configs" ) ]]; then
+		rm -rf "lib/ul_${libs[$i]}"
+		mv "tmp/lib/ul_${libs[$i]}" lib
+	fi
 done
 
 rm -rf tmp
